test(groups): Move tests to separate module
- Move related tests to separate module
- Make config dependency explicit in `build.rs`
